Organize Rust SDK by user concepts
