+++ sdk-rust/crates/dex-sdk/src/lib.rs
@@ -6,27 +6,55 @@
 //
 // SPDX-License-Identifier: LicenseRef-Super-Durable-1.0
 
+mod attribute;
+mod channel;
 mod client;
+mod client_options;
+mod context;
 mod error;
 mod flow;
-mod options;
+mod flow_config;
+mod flow_info;
+mod persistence;
+mod registry;
+mod reset_flow_options;
+mod retry_policy;
 mod rpc;
-mod state;
+mod start_flow_options;
+mod step;
+mod step_execution;
+mod step_options;
+mod stop_flow_options;
+mod timer;
+mod value;
 mod wait;
+mod worker;
+mod worker_options;
 
-pub use client::{
-    Client, ClientOptions, FlowErrorType, FlowInfo, FlowStatus, Registry, Worker, WorkerOptions,
-};
+pub use attribute::{Attribute, AttributeIndex, AttributeMap};
+pub use channel::{Channel, ChannelMap};
+pub use client::Client;
+pub use client_options::ClientOptions;
+pub use context::Context;
 pub use dex_blob_cache::{BlobCache, BlobCacheConfig};
 pub use error::{HandlerError, HandlerResult, SdkError, SdkResult};
-pub use flow::{Flow, Step, StepDecision, StepExecutionId, StepList, StepMovement, Value};
-pub use options::{
-    ActiveStepSearchMode, FlowConfig, IdReusePolicy, ResetFlowOptions, RetryPolicy,
-    StartFlowOptions, StepDurability, StepOptions, StopFlowOptions, WaitForFailurePolicy,
-    WorkerTarget,
-};
+pub use flow::Flow;
+pub use flow_config::{ActiveStepSearchMode, FlowConfig};
+pub use flow_info::{FlowErrorType, FlowInfo, FlowStatus};
+pub use persistence::PersistenceSchema;
+pub use registry::Registry;
+pub use reset_flow_options::ResetFlowOptions;
+pub use retry_policy::RetryPolicy;
 pub use rpc::{Rpc, RpcList, RpcResult};
-pub use state::{
-    Attribute, AttributeIndex, AttributeMap, Channel, ChannelMap, Context, PersistenceSchema,
-};
-pub use wait::{Condition, ConditionCombination, Timer, TimerId, Wait};
+pub use start_flow_options::{IdReusePolicy, StartFlowOptions};
+pub use step::{Step, StepDecision, StepList, StepMovement};
+pub use step_execution::StepExecutionId;
+pub use step_options::{StepDurability, StepOptions, WaitForFailurePolicy};
+pub use stop_flow_options::StopFlowOptions;
+pub use timer::{Timer, TimerId};
+pub use value::Value;
+pub use wait::{Condition, ConditionCombination, Wait};
+pub use worker::Worker;
+pub use worker_options::{WorkerOptions, WorkerTarget};
+
+pub(crate) use value::short_type_name;
