+++ sdk-rust/crates/dex-sdk/src/persistence.rs
@@ -1,0 +1,48 @@
+// Copyright (c) 2026 Super Durable, Inc.
+//
+// Licensed under the Super Durable Source License 1.0.
+// You may not use this file except in compliance with the License.
+// See the LICENSE file in the repository root.
+//
+// SPDX-License-Identifier: LicenseRef-Super-Durable-1.0
+
+use crate::{Attribute, AttributeMap, Channel, ChannelMap};
+
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
+pub struct PersistenceSchema {
+    definitions: Vec<PersistenceDefinition>,
+}
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+enum PersistenceDefinition {
+    Attribute,
+    AttributeMap,
+    Channel,
+    ChannelMap,
+}
+
+impl PersistenceSchema {
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn attribute<T>(mut self, _attribute: &Attribute<T>) -> Self {
+        self.definitions.push(PersistenceDefinition::Attribute);
+        self
+    }
+
+    pub fn attribute_map<T>(mut self, _attribute: &AttributeMap<T>) -> Self {
+        self.definitions.push(PersistenceDefinition::AttributeMap);
+        self
+    }
+
+    pub fn channel<T>(mut self, _channel: &Channel<T>) -> Self {
+        self.definitions.push(PersistenceDefinition::Channel);
+        self
+    }
+
+    pub fn channel_map<T>(mut self, _channel: &ChannelMap<T>) -> Self {
+        self.definitions.push(PersistenceDefinition::ChannelMap);
+        self
+    }
+}
